requirements.txt
discord.py
openai
nest_asyncio
